Fix of first review pass.

+++ source/gameengine/Ketsji/KX_KetsjiEngine.h
@@ -238,8 +238,8 @@ private:
 	 */
 	void UpdateSuspendedScenes();
 
-	void RenderFrame(KX_Scene *scene, KX_Camera *cam, int pass);
-	void PostRenderScene(KX_Scene *scene, int target);
+	void RenderFrame(KX_Scene *scene, KX_Camera *cam, unsigned short pass);
+	void PostRenderScene(KX_Scene *scene, unsigned short target);
 	void RenderDebugProperties();
 
 public:
